Assuming same origin and the need to pass creds
This change now assumes that AJAX requests will be going to the same host (and
port) as the JavaScript. This could be because Echo is serving the JavaScript
as a static asset, or nginx is serving it.

Also, credentials are now passed along with the AJAX request. This in necessary
when running behind the argonaut-gateway.